franchise_scrapers/requirements.txt
# franchise_scrapers requirements
playwright>=1.40.0
pydantic>=2.0.0
typer>=0.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
rich>=13.0.0